fix: change update method to use id and change the dao class to accept long instead of int

+++ src/main/java/repository/AuthorDao.java
@@ -16,7 +16,7 @@ public class AuthorDao {
         this.jdbcTemplate = jdbcTemplate;
     }
 
-    public Author findById(Integer id) {
+    public Author findById(Long id) {
         String sql = "SELECT * FROM author WHERE id = ?";
 
         return jdbcTemplate.queryForObject(sql, rowMapper, id);
